Major Project (1920*1080 version)
The COMP1720 major project is to create an interactive artwork for a new media art installation. Gallery attendees are able to walk around and observe the various works (including yours) at their leisure. If they wish, they can pause at your sketch and interact with it, but they will receive no additional guidance/instruction on what to do. Your goal is to provide an engaging user experience of roughly three minutes, but the exact nature of that experience is up to you. It might be an interactive generative artwork, it might be an interactive movie/storytelling experience, it might be an arty game—you get to choose.